Project2/requirements.txt
# 数字水印系统依赖包

# 核心图像处理库
opencv-python>=4.5.0
numpy>=1.19.0
pillow>=8.0.0

# 可视化和绘图
matplotlib>=3.3.0

# 测试框架
pytest>=6.0.0
pytest-cov>=2.10.0

# 代码格式化和检查
black>=21.0.0

# 可选：开发工具
# flake8>=3.8.0  # 已禁用lint检查 